pyproject.toml
[tool.poetry]
name = "iambic"
version = "3.0.0"
description = "Data extraction and rendering library for Shakespearean text."
authors = ["Sean Stewart <sean_stewart@me.com>"]
license = "MIT"
readme = "README.md"
repository = "https://github.com/seandstewart/iambic"
keywords = ["text", "render", "shakespeare", "text-processing"]
classifiers = [
    "Development Status :: 5 - Production/Stable",
    "Intended Audience :: Developers",
    "Natural Language :: English",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Programming Language :: Python",
    "Programming Language :: Python :: 3.6",
    "Programming Language :: Python :: 3.7",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3 :: Only",
    "Programming Language :: Python :: Implementation :: CPython",
    "Topic :: Utilities",
    "Topic :: Text Processing",
    "Topic :: Text Processing :: Indexing",
    "Topic :: Text Processing :: Linguistic",
    "Topic :: Text Processing :: Markup",
    "Topic :: Text Processing :: Markup :: HTML",
    "Typing :: Typed",
]
include = ["*.md", "*.json"]

[tool.poetry.dependencies]
python = "^3.11"
typical = "^2"
ujson = "^2"
markdown = "^3"
pymdown-extensions = "^9"
html2text = "^2020.1.16"
tabulate = "^0.8.9"

[tool.poetry.group.docs.dependencies]
mkdocs-awesome-pages-plugin = "^2"
mkdocs = "^1"
mkdocs-material = "^9"
fontawesome-markdown = "^0.2.6"

[tool.poetry.group.dev.dependencies]
pytest = "^7"
pytest-cov = "^4"
pre-commit = "^3"
flake8 = "^6"
flake8-bugbear = "^23"
mypy = "^1"
black = "^23"
isort = "^5"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"

[tool.mypy]
mypy_path = "$MYPY_CONFIG_FILE_DIR/"
no_namespace_packages = true
python_version = "3.11"
install_types = true
non_interactive = true
ignore_missing_imports = true
no_strict_optional = true
follow_imports = "silent"
exclude = ".*tests/.*|.*docs/.*"
